+++ chiya/cogs/commands/general.py
@@ -3,7 +3,7 @@ import logging
 import discord
 from chiya import config
 from chiya.utils import embeds
-from discord.commands import Option, context, slash_command, message_command
+from discord.commands import Option, context, slash_command
 from discord.ext import commands
 import discord.utils
 
@@ -54,39 +54,6 @@ class GeneralCommands(commands.Cog):
             embed.set_author(icon_url=user.display_avatar, name=str(user))
             embed.set_image(url=user.display_avatar)
         await ctx.send_followup(embed=embed)
-
-    @message_command(
-        guild_ids=config["guild_ids"],
-        description="Add vote reactions to a message.",
-    )
-    @commands.has_role(config["roles"]["staff"])
-    async def vote(
-        self,
-        ctx: context.ApplicationContext,
-        message: discord.Message,
-    ) -> None:
-        """
-        Adds vote emojis (yes and no) reactions to a message.
-        """
-        # TODO: what happens if the user doesn't have permission to add reactions in that channel?
-        await ctx.defer(ephemeral=True)
-
-        if message:
-            try:
-                message = await ctx.channel.fetch_message(message)
-            except discord.NotFound:
-                return await embeds.error_message(ctx=ctx, description="Invalid message ID.")
-
-        if not message:
-            messages = await ctx.channel.history(limit=1).flatten()
-            message = messages[0]
-
-        emoji_yes = discord.utils.get(ctx.guild.emojis, id=config['emoji']['yes']) or "👍"
-        emoji_no = discord.utils.get(ctx.guild.emojis, id=config['emoji']['no']) or "👎"   
-
-        await message.add_reaction(emoji_yes)
-        await message.add_reaction(emoji_no)
-        await embeds.success_message(ctx=ctx, description=f"Added votes to {message.jump_url}")
 
     @slash_command(
         guild_ids=config["guild_ids"],
